significantly changed polyglot suite

+++ evaluator/datasets/polyglot/book-store/main.py
@@ -1,2 +1,2 @@
-def total(basket: list[int]) -> float: # in cents
+def total(basket: list[int]) -> int: # in cents
     pass
